backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
telethon==1.37.0
python-docx==1.1.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
pyjwt==2.9.0
cryptography==43.0.3
supabase==2.9.1
